#15401 tests are blind - have not been run (#15402)
* #15401 tests are blind - have not been run

* #15401 allow change template of a page is template_version_info is deleted

* #15401 fail test when NPE

+++ dotCMS/src/main/java/com/dotmarketing/portlets/contentlet/ajax/ContentletAjax.java
@@ -1582,7 +1582,7 @@ public class ContentletAjax {
 						final Template previousTemplate = APILocator.getTemplateAPI().findWorkingTemplate(previousTemplateId,
 								systemUser, false);
 
-						if (previousTemplate.isAnonymous()) {
+						if (UtilMethods.isSet(previousTemplate) && previousTemplate.isAnonymous()) {
 							APILocator.getTemplateAPI().delete(previousTemplate, systemUser, false);
 						}
 					}
